split controller_api into controller_{api,protocol}
