+++ gateserver/controller_api.py
@@ -1,80 +1,15 @@
-from .utils.structparse import *
-from . import utils
+from .controller_protocol import *
 from . import db
-import nacl.raw as nacl
-
-PROTOCOL_VERSION = bytes([0x00,0x01])
-
-class BadMessageError(Exception): pass
-
-def checkmsg(expression, err):
-    if not expression: raise BadMessageError(err)
-
-class MsgType(Enum):
-    OPEN = 1
-
-class ReplyStatus(Enum):
-    OK = 0x01
-    ERR = 0x10
-    TRY_AGAIN = 0x11
-S = ReplyStatus
-
-PacketHead = mystruct('PacketHead',
-                     ['protocol_version', 'controllerID', 'nonce' ],
-                     [ t.bytes(2) , t.bytes(6) , t.bytes(18) ])
-
-RequestHead = mystruct('RequestHead',
-                      ['msg_type'],
-                      [ t.uint8 ])
-
-ReplyHead = mystruct('ReplyHead',
-                    ['msg_type', 'status' ],
-                    [ t.uint8 , t.uint8 ])
+from . import utils
 
 def get_key_for_mac(mac):
     rs = db.exec_sql('SELECT key FROM controller WHERE id = %s',
                      (utils.bytes2mac(mac),), ret=True)
-    checkmsg(len(rs) == 1, 'unknown controllerID '+utils.bytes2mac(mac))
+    checkmsg(len(rs) == 1, 'unknown controllerID ')
     return rs[0]['key'].tobytes()
 
-def crypto_unwrap(packet_head, payload):
-    key = get_key_for_mac(packet_head.controllerID)
-    return nacl.crypto_secretbox_open(payload,
-        packet_head.controllerID + packet_head.nonce, key)
-
-def crypto_wrap(packet_head, payload):
-    key = get_key_for_mac(packet_head.controllerID)
-    return nacl.crypto_secretbox(payload,
-        packet_head.controllerID + packet_head.nonce, key)
-
-def parse_packet(buf, struct=None):
-    struct = struct or RequestHead
-    assert struct in [RequestHead, ReplyHead]
-    p, payload = PacketHead.unpack_from(buf)
-    checkmsg(p.protocol_version == PROTOCOL_VERSION, 'Invalid protocol version')
-    payload = crypto_unwrap(p, payload)
-    r, data = struct.unpack_from(payload)
-    try:
-        t = MsgType(r.msg_type)
-    except ValueError:
-        raise BadMessageError('Unknown request type {}'.format(r.msg_type))
-    return p, r, t, data
-
-def make_packet(packet_head, r_head, data=None):
-    """Requires `packet_head` and `r_head` to be valid."""
-    payload = r_head.pack() + (data or bytes(0))
-    return packet_head.pack() + crypto_wrap(packet_head, payload)
-
-def make_reply_for(packet_head, request_head, status, data=None):
-    """Requires `packet_head` and `request_head` to be valid."""
-    nnonce = bytearray(packet_head.nonce); nnonce[-1] ^= 0x1
-    p = PacketHead(protocol_version=PROTOCOL_VERSION,
-                   controllerID=packet_head.controllerID,
-                   nonce=nnonce)
-    r = ReplyHead(msg_type=request_head.msg_type, status=status.value)
-    return make_packet(p, r, data)
-
 def fstring(buf):
+    """See https://github.com/fmfi-svt-gate/server/wiki/Controller-%E2%86%94-Server-Protocol#note-isic-ids-representation"""
     length, string = int(buf[0]), buf[1:]
     checkmsg(length <= len(string), 'fstring length > buffer size')
     return string[:length]
@@ -85,7 +20,21 @@ process_request = {
 }
 assert set(MsgType) == set(process_request), 'Not all message types are handled'
 
+def log_message(controllerID, req_type, indata, status):
+    """TODO"""
+    print(utils.bytes2mac(controllerID), req_type.name, indata, '->', status.name)
+
+def log_bad_packet(buf, e):
+    """TODO"""
+    raise e
+
 def handle_request(buf):
-    packet_head, request_head, request_type, indata = parse_packet(buf)
-    status, outdata = process_request[request_type](indata)
-    return make_reply_for(packet_head, request_head, status, outdata)
+    try:
+        packet_head, payload = parse_packet_head(buf)
+        key = get_key_for_mac(packet_head.controllerID)
+        req_head, req_type, indata = parse_r(RequestHead, packet_head, key, payload)
+        status, outdata = process_request[req_type](indata)
+        log_message(packet_head.controllerID, req_type, indata, status)
+        return make_reply_for(packet_head, req_head, key, status, outdata)
+    except BadMessageError as e:
+        log_bad_packet(buf, e)
